rename customscaffold to appscaffold
