+++ lib/widgets/pages/receipts/receipts_page.dart
@@ -1,8 +1,8 @@
 import 'package:coffeecard/base/strings.dart';
+import 'package:coffeecard/widgets/app_scaffold.dart';
 import 'package:coffeecard/widgets/components/receipt/filter_bar.dart';
 import 'package:coffeecard/widgets/components/receipt/receipt_dropdown.dart';
 import 'package:coffeecard/widgets/components/receipt/receipts_list_view.dart';
-import 'package:coffeecard/widgets/custom_scaffold.dart';
 import 'package:flutter/material.dart';
 
 class ReceiptsPage extends StatelessWidget {
