+++ lib/widgets/pages/redeem_voucher_page.dart
@@ -1,5 +1,5 @@
 import 'package:coffeecard/base/strings.dart';
-import 'package:coffeecard/widgets/custom_scaffold.dart';
+import 'package:coffeecard/widgets/app_scaffold.dart';
 import 'package:flutter/material.dart';
 
 class RedeemVoucherPage extends StatelessWidget {
