+++ lib/widgets/pages/tickets_page.dart
@@ -1,8 +1,8 @@
 import 'package:coffeecard/base/strings.dart';
+import 'package:coffeecard/widgets/app_scaffold.dart';
 import 'package:coffeecard/widgets/components/analog_closed_popup.dart';
 import 'package:coffeecard/widgets/components/shop_section.dart';
 import 'package:coffeecard/widgets/components/tickets_section.dart';
-import 'package:coffeecard/widgets/custom_scaffold.dart';
 import 'package:flutter/material.dart';
 
 class TicketsPage extends StatefulWidget {
